Attempt to function with FF.
I've done a bit of researching and hacking about to see if I could make a
Firefox Add-on version.

It's displaying the OP button and the comment works. I've also manually copied
the browser-poyfill.js file from the `node_modules/webextension-polyfill/dist/`
directory.

At this stage, it's all about webextension-polyfill trying to make things
cross-browser compatible, but quite a bit more needs to be done to get it
working and done professionally.

And... it seems to have borked the commenting option in Chrome. Grrrr...